requirements.txt
langchain
langchain-community
langchain-groq
unstructured
selenium
chromadb
streamlit
pandas
python-dotenv
pyperclip
streamlit-extras